feat: Introduce DtoNamingHelper for centralized DTO name generation and refactor related methods
