+++ src/Linqraft.Analyzer/SelectExprToTypedCodeFixProvider.cs
@@ -192,94 +192,6 @@ public class SelectExprToTypedCodeFixProvider : CodeFixProvider
         AnonymousObjectCreationExpressionSyntax anonymousType
     )
     {
-        string baseName;
-
-        // Walk up the tree to find the relevant context
-        var current = invocation.Parent;
-        while (current != null)
-        {
-            switch (current)
-            {
-                // Check for variable declaration: var name = query.SelectExpr(...)
-                case EqualsValueClauseSyntax equalsValue
-                    when equalsValue.Parent is VariableDeclaratorSyntax declarator:
-                    var varName = declarator.Identifier.Text;
-                    baseName = ToPascalCase(varName) + "Dto";
-                    return baseName + "_" + GenerateHash(anonymousType);
-
-                // Check for assignment: name = query.SelectExpr(...)
-                case AssignmentExpressionSyntax assignment
-                    when assignment.Left is IdentifierNameSyntax identifier:
-                    baseName = ToPascalCase(identifier.Identifier.Text) + "Dto";
-                    return baseName + "_" + GenerateHash(anonymousType);
-
-                // Stop at statement level
-                case StatementSyntax:
-                    break;
-
-                default:
-                    current = current.Parent;
-                    continue;
-            }
-            break;
-        }
-
-        // Check for return statement with method name
-        var methodDecl = invocation.Ancestors().OfType<MethodDeclarationSyntax>().FirstOrDefault();
-        if (methodDecl != null)
-        {
-            var methodName = methodDecl.Identifier.Text;
-            if (methodName.StartsWith("Get"))
-            {
-                methodName = methodName.Substring(3);
-            }
-            baseName = methodName + "Dto";
-            return baseName + "_" + GenerateHash(anonymousType);
-        }
-
-        return "ResultDto_" + GenerateHash(anonymousType);
-    }
-
-    private static string GenerateHash(AnonymousObjectCreationExpressionSyntax anonymousType)
-    {
-        // Generate hash based on property names
-        var sb = new StringBuilder();
-        foreach (var initializer in anonymousType.Initializers)
-        {
-            string propertyName;
-            if (initializer.NameEquals != null)
-            {
-                propertyName = initializer.NameEquals.Name.Identifier.Text;
-            }
-            else
-            {
-                propertyName = GetPropertyNameFromExpression(initializer.Expression);
-            }
-            sb.Append(propertyName);
-            sb.Append(';');
-        }
-
-        // Use common hash utility
-        return HashUtility.GenerateAlphanumericHash(sb.ToString());
-    }
-
-    private static string GetPropertyNameFromExpression(ExpressionSyntax expression)
-    {
-        return expression switch
-        {
-            MemberAccessExpressionSyntax memberAccess => memberAccess.Name.Identifier.Text,
-            IdentifierNameSyntax identifier => identifier.Identifier.Text,
-            ConditionalAccessExpressionSyntax conditionalAccess
-                when conditionalAccess.WhenNotNull is MemberBindingExpressionSyntax memberBinding =>
-                memberBinding.Name.Identifier.Text,
-            _ => "Property",
-        };
-    }
-
-    private static string ToPascalCase(string name)
-    {
-        if (string.IsNullOrEmpty(name))
-            return name;
-        return char.ToUpperInvariant(name[0]) + name.Substring(1);
+        return DtoNamingHelper.GenerateDtoName(invocation, anonymousType);
     }
 }
